feat: Added (very barebones) MainMenu Scene.
It utilizes a temporary background and only has one button. It is configured to scale with the screen resolution. The camera has a background color of white. The button has all its states (for color) set to zero alpha (transparent). The MainMenuGroup Empty Object is merely for organization/cleanliness and is in no way functional.